chore/idata_handler: add new duplicate msg b/w nodes for chunk duplication
